% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_package.R
\name{build.package}
\alias{build.package}
\title{Build R-packages with devtools}
\usage{
build.package(directory)
}
\description{
Build R-packages with devtools
}
\details{
\code{Builds Rd files and updates description and Namespace}.
}
\examples{
\donttest{# build.package("/Path/to/dir/packagename/R")
}
}
