% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create.package}
\alias{create.package}
\title{Create R-packages with devtools}
\usage{
create.package(directory)
}
\description{
Create R-packages with devtools
}
\details{
\code{Creates all necessary files and directories for a package}.
}
\examples{
\donttest{# create.package("/Path/to/dir/packagename")
}
}
