% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get.tables}
\alias{get.tables}
\title{Get tables of sqlite3 databases}
\usage{
get.tables(data.base)
}
\description{
Get tables of sqlite3 databases
}
\details{
\code{Lists all available tables of a databse}.
}
\examples{
\donttest{# List tables of test.sqlite3
tables=get.tables("test.sqlite3")
}
}
