% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_anc.R
\name{remove.anc}
\alias{remove.anc}
\alias{remove_anc}
\title{Remove symbols from a string not allowed for filenames and folders}
\usage{
remove.anc(string)
}
\description{
Remove symbols from a string not allowed for filenames and folders
}
\details{
\code{Deletes \\ / : * ?  < > |}.
}
\examples{
\donttest{# string="?a></lex"
remove.anc(string)
}
}
