% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_NAcols.R
\name{remove.NAcols}
\alias{remove.NAcols}
\title{Delete columns of a dataframe containing NAs}
\usage{
remove.NAcols(dframe)
}
\description{
Delete columns of a dataframe containing NAs
}
\details{
\code{Remove all columns from a dataframe containing NAs (na.omit excludes rows)}.
}
\examples{
\donttest{# dataframe=remove.NAcols(dataframe)
}
}
