% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cutoffs.R
\name{setCutoff}
\alias{setCutoff}
\title{Plot Density Histograms and set cutoffs with mouse}
\usage{
setCutoff(exprData, dir = getwd(), style = "h", color = "steelblue3", ...)
}
\description{
Plot Density Histograms and set cutoffs with mouse
}
\details{
\code{Plots density histograms and saves cutoffs in a dataframe}.
}
\examples{
\donttest{# setCutof(expressionData)
}
}
