% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDB.R
\name{writeDB}
\alias{writeDB}
\title{Write or update a table into an existing or not existing (then it will be generated)
database}
\usage{
writeDB(file, table, df, add = F, replace = T, verbose = F)
}
\description{
Write or update a table into an existing or not existing (then it will be generated)
database
}
\details{
\code{This Function needs a .sqlite3 databasefile, title of the table to 
write and a dataframe to store. It is possible to append data to an existing table
and to overwrite it. The verbose flag is used to display work progress to console}.
}
\examples{
\donttest{# writeDB("FlowRepExperiments.sqlite3",
"Experiment_table_short",experiments,verbose=T)
}
}
