% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_FlowRep.R
\name{updateFlowRep}
\alias{updateFlowRep}
\alias{upflo}
\title{Update FlowRepository datasets}
\usage{
updateFlowRep(verbose = F)
}
\description{
Update FlowRepository datasets
}
\details{
\code{Special function for PRIbase, updates FlowRepository datasets
and returns TRUE if new experiments are available else FALSE}.
}
\examples{
\donttest{# updateFlowRep(T) # displays info in console
}
}
